+++ data-ingestion-service/src/main/java/gov/cdc/dataingestion/config/LiquibaseConfig.java
@@ -14,41 +14,69 @@ import javax.sql.DataSource;
 @Configuration
 public class LiquibaseConfig {
 
+    @Value("${spring.liquibase.driver-class-name}")
+    private String driverClassName;
+
     @Value("${spring.liquibase.user}")
     private String dbUserName;
 
     @Value("${spring.liquibase.password}")
     private String dbUserPassword;
 
-    @Value("${spring.liquibase.driver-class-name}")
-    private String driverClassName;
-
-    @Value("${spring.liquibase.url.dataingest}")
-    private String dbUrl;
-
     @Bean
-    @ConfigurationProperties(prefix = "spring.liquibase")
-    public LiquibaseProperties liquibaseProperties() {
+    @ConfigurationProperties(prefix = "spring.liquibase.dataingest")
+    public LiquibaseProperties dataingestLiquibaseProperties() {
         return new LiquibaseProperties();
     }
 
-    @Bean(name = "liquibaseDataSource")
-    public DataSource liquibaseDataSource() {
+    @Bean
+    public DataSource dataingestDataSource(@Qualifier("dataingestLiquibaseProperties") LiquibaseProperties props) {
         return DataSourceBuilder.create()
-                .url(dbUrl)
+                .url(props.getUrl())
                 .username(dbUserName)
                 .password(dbUserPassword)
                 .driverClassName(driverClassName)
                 .build();
     }
 
-    @Bean
-    public SpringLiquibase liquibase(@Qualifier("liquibaseDataSource") DataSource liquibaseDataSource, LiquibaseProperties liquibaseProperties) {
+    @Bean(name = "dataingestLiquibase")
+    public SpringLiquibase dataingestLiquibase(
+            @Qualifier("dataingestDataSource") DataSource dataSource,
+            @Qualifier("dataingestLiquibaseProperties") LiquibaseProperties props) {
         SpringLiquibase liquibase = new SpringLiquibase();
-        liquibase.setDataSource(liquibaseDataSource);
-        liquibase.setChangeLog(liquibaseProperties.getChangeLog());
-        if (liquibaseProperties.getContexts() != null) {
-            liquibase.setContexts(String.join(",", liquibaseProperties.getContexts()));
+        liquibase.setDataSource(dataSource);
+        liquibase.setChangeLog(props.getChangeLog());
+        if (liquibase.getContexts() != null) {
+            liquibase.setContexts(String.join(",", liquibase.getContexts()));
+        }
+        return liquibase;
+    }
+
+    @Bean
+    @ConfigurationProperties(prefix = "spring.liquibase.msgoute")
+    public LiquibaseProperties msgouteLiquibaseProperties() {
+        return new LiquibaseProperties();
+    }
+
+    @Bean
+    public DataSource msgouteDataSource(@Qualifier("msgouteLiquibaseProperties") LiquibaseProperties props) {
+        return DataSourceBuilder.create()
+                .url(props.getUrl())
+                .username(dbUserName)
+                .password(dbUserPassword)
+                .driverClassName(driverClassName)
+                .build();
+    }
+
+    @Bean(name = "msgouteLiquibase")
+    public SpringLiquibase msgouteLiquibase(
+            @Qualifier("msgouteDataSource") DataSource dataSource,
+            @Qualifier("msgouteLiquibaseProperties") LiquibaseProperties props) {
+        SpringLiquibase liquibase = new SpringLiquibase();
+        liquibase.setDataSource(dataSource);
+        liquibase.setChangeLog(props.getChangeLog());
+        if (liquibase.getContexts() != null) {
+            liquibase.setContexts(String.join(",", liquibase.getContexts()));
         }
         return liquibase;
     }
