+++ data-ingestion-service/src/main/resources/db/dataingest/create-nbs-dataingest-db.sql
