+++ data-ingestion-service/src/main/resources/db/dataingest/tables/di-service-001.sql
